Add picture macNcheese, Delete picture beef-stew, Have recipe lasagne
TOP changed their recipe example site.
I was to lazy to find a recipe myself.
I changed beef-stew to lasagne.
I thought the site was a recipe for lasagne.
Turns out - it is a list of recipes for casserolles and macNcheeses.
To late to change everything now.
Need to change "lasagne" - > 'macNcheese'